Refactor: Remove bussiness logic from Canvas
